server/requirement.txt
uvicorn
fastapi
python-dotenv
openai
git+https://github.com/openai/whisper.git
ffmpeg
requests